Add color to header
